Integration tests: separate test from initramfs
Our current test infra builds the test itself into the initramfs, which
means that we cannot separating building and testing or test with a
custom initramfs. To solve this, we separate the test from the
initramfs as follows:

Instead of building a uinit.go with the caller's provided commands, have
a generic uinit.go that finds an Elvish shell script of test commands in
the P9 directory exposed to the VM. This way callers can either 1)
provide a custom uinit or 2) provide commands to run for testing.

As part of that, we add some options to vmtest.Options and change the
tmpDir to be easier to work with.

Signed-off-by: pallaud <plaud@google.com>